refactor: split ui and combat modules
