+++ modules/gameState.js
@@ -1,0 +1,9 @@
+export const game = {
+  player: null,
+  target: null,
+  combatTimer: 0,
+  inCombat: false,
+  onlinePlayers: [],
+  players: {},
+  currentZone: { id: null, mobs: [] }
+};
